+++ routes/web.php
@@ -11,6 +11,11 @@
 |
 */
 
+Route::get('/', function () {
+    return view('welcome');
+});
+
+
 Auth::routes(['register' => false]);
 
 Route::get('/home', 'HomeController@index')->name('home');