+++ app/Http/Controllers/RatesAPIV2Controller.php
@@ -14,14 +14,14 @@ class RatesAPIV2Controller extends Controller
 
     private $gstFields;
     private $pstFields;
+    private $provinces_codes;
 
     public function __construct()
     {
+        $this->provinces_codes = ['ab', 'bc', 'mb', 'nl', 'ns', 'nt', 'nu', 'on', 'pe', 'qc', 'sk', 'yt'];
         $this->gstFields = ['start', 'type', 'gst', 'applicable', 'source', 'updated_at'];
         $this->pstFields = ['start', 'type', 'pst', 'hst', 'gst', 'applicable', 'source', 'updated_at'];
         $this->allPstFields = ['province', 'start', 'type', 'pst', 'hst', 'gst', 'applicable', 'source', 'updated_at'];
-
-        dispatch(new IncrementStats(request()->path()));
     }
 
     /**
@@ -29,6 +29,8 @@ class RatesAPIV2Controller extends Controller
      */
     public function getCurrentGst()
     {
+        $this->incrementStats();
+
         return Cache::remember('gst-current-rate', 86400, function () {
             $rates = Rates::where('province', 'all')
                     ->where('start', '<=', Carbon::now())
@@ -57,6 +59,8 @@ class RatesAPIV2Controller extends Controller
      */
     public function getFutureGst()
     {
+
+        $this->incrementStats();
         return Cache::remember('gst-future-rate', 86400, function () {
             $rate = Rates::where('province', 'all')
                     ->where('start', '>', Carbon::now())
@@ -76,6 +80,8 @@ class RatesAPIV2Controller extends Controller
      */
     public function getHistoricalGst()
     {
+        $this->incrementStats();
+
         return Cache::remember('gst-all-rates', 86400, function () {
             return Rates::where('province', 'all')
                     ->orderBy('start', 'DESC')
@@ -88,6 +94,7 @@ class RatesAPIV2Controller extends Controller
      */
     public function getAllPst()
     {
+        $this->incrementStats();
 
         return Cache::remember('pst-all-current-rate', 86400, function () {
             $all = Rates::where('province', '!=', 'all')
@@ -114,6 +121,10 @@ class RatesAPIV2Controller extends Controller
      */
     public function getCurrentPst($province)
     {
+        $this->checkProvinceCodeValidity($province);
+
+        $this->incrementStats();
+        
         return Cache::remember("pst-{$province}-current-rate", 86400, function () use ($province) {
             $rates = Rates::where('province', $province)
                     ->where('start', '<=', Carbon::now())
@@ -143,6 +154,10 @@ class RatesAPIV2Controller extends Controller
      */
     public function getFuturePst($province)
     {
+        $this->checkProvinceCodeValidity($province);
+
+        $this->incrementStats();
+
         return Cache::remember("pst-{$province}-future-rate", 86400, function () use ($province) {
             $rate = Rates::where('province', $province)
                     ->where('start', '>', Carbon::now())
@@ -163,10 +178,30 @@ class RatesAPIV2Controller extends Controller
      */
     public function getHistoricalPst($province)
     {
+        $this->checkProvinceCodeValidity($province);
+
+        $this->incrementStats();
+        
         return Cache::remember("{$province}-all-rates", 86400, function () use ($province) {
             return Rates::where('province', $province)
                     ->orderBy('start', 'DESC')
                     ->get($this->pstFields);
         });
+    }
+
+    private function incrementStats()
+    {
+        dispatch(new IncrementStats(request()->path()));
+    }
+
+    private function checkProvinceCodeValidity($code)
+    {
+        $province_code_is_valid = in_array($code, $this->provinces_codes) ? true : false;
+
+        if (!$province_code_is_valid) {
+            abort(404, 'Invalid two letter province code.');
+        }
+
+        return true;
     }
 }