+++ changelog.md
@@ -4,6 +4,7 @@ Trying as hard as I can to stick to semantic versionning.
 
 ## Next
 - Added an index page at the root of api.salestaxapi.ca
+- Error logging is set to daily rather than single file.
 
 ## 4.4.0 (2019-06-01)
 - Removed ability to register from the frontend.